fix: added the default general configuration on application start up
